backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
torch==2.1.0
transformers==4.35.0
peft==0.7.0
pytesseract
Pillow
groq
python-multipart
